Moved chatbot to seperate folder
