+++ app.py
@@ -1,6 +1,6 @@
 from flask import Flask, render_template , jsonify,request
 
-from chat import get_response
+from chatbot.chat import get_response
  
 
 app=Flask(__name__)
